+++ src/core/operations/BitwiseOp.js
@@ -1,5 +1,4 @@
 import Utils from "../Utils.js";
-import CryptoJS from "crypto-js";
 
 
 /**
