+++ src/core/operations/DateTime.js
@@ -1,5 +1,3 @@
-import {BigInteger} from "jsbn";
-
 /**
  * Date and time operations.
  *
@@ -78,85 +76,6 @@ const DateTime = {
             throw "Unrecognised unit";
         }
     },
-
-
-    /**
-     * Windows Filetime to Unix Timestamp operation.
-     *
-     * @author bwhitn [brian.m.whitney@outlook.com]
-     * @param {string} input
-     * @param {Object[]} args
-     * @returns {string}
-     */
-    runFromFiletimeToUnix: function(input, args) {
-        let units = args[0],
-            format = args[1];
-
-        if (format === "Hex") {
-            input = new BigInteger(input, 16);
-        } else {
-            input = new BigInteger(input);
-        }
-
-        input = input.subtract(new BigInteger("116444736000000000"));
-
-        if (units === "Seconds (s)"){
-            input = input.divide(new BigInteger("10000000"));
-        } else if (units === "Milliseconds (ms)") {
-            input = input.divide(new BigInteger("10000"));
-        } else if (units === "Microseconds (μs)") {
-            input = input.divide(new BigInteger("10"));
-        } else if (units === "Nanoseconds (ns)") {
-            input = input.multiply(new BigInteger("100"));
-        } else {
-            throw "Unrecognised unit";
-        }
-
-        return input.toString();
-    },
-
-
-    /**
-     * Unix Timestamp to Windows Filetime operation.
-     *
-     * @author bwhitn [brian.m.whitney@outlook.com]
-     * @param {string} input
-     * @param {Object[]} args
-     * @returns {string}
-     */
-    runToFiletimeFromUnix: function(input, args) {
-        let units = args[0],
-            format = args[1];
-
-        input = new BigInteger(input);
-
-        if (units === "Seconds (s)"){
-            input = input.multiply(new BigInteger("10000000"));
-        } else if (units === "Milliseconds (ms)") {
-            input = input.multiply(new BigInteger("10000"));
-        } else if (units === "Microseconds (μs)") {
-            input = input.multiply(new BigInteger("10"));
-        } else if (units === "Nanoseconds (ns)") {
-            input = input.divide(new BigInteger("100"));
-        } else {
-            throw "Unrecognised unit";
-        }
-
-        input = input.add(new BigInteger("116444736000000000"));
-
-        if (format === "Hex"){
-            return input.toString(16);
-        } else {
-            return input.toString();
-        }
-    },
-
-
-    /**
-     * @constant
-     * @default
-     */
-    FILETIME_FORMATS: ["Decimal", "Hex"],
 
 
     /**
